+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 [*] other changes
 
 0.9.8
-- [+] add country rank info for each song
+- [+] show country rank info for each song
+- [+] add country rank filtering, thx Rocker
 
 0.9.7.9
 - [*] parsing PPs on the country leaderboard in addition to API data, thx Rocker